libwimax: move to address interfaces by attribute (vs by family)
Commnunity review asked to move the whole genl schema to addess
interfaces by ifindex contained in an attribute on each message.

Doing so made it almost impossible to keep the models we had with a
single family per interface, so huge rework ensued:

 - remove pipes and all their APIs: no longer makes sense to have one
   mcast group per pipe, so this is no longer needed.

 - create the pipe concept by adding a "pipe name" to the
   message. Default pipe has no name; as a side benefit, all the pipes
   are now bidirectional.

 - side benefit: greatly simplify the msg vs pipe vs mc_rx interface:
   there is only one interface now, wimax_msg_*() and it contains just
   a few functions.

 - Delete all the usage of multicast groups and multicast groups
   handles

 - Declare two handles, one for reading notifications, one for
   writing (eases the work)

All in all, this cuts a lot of code; on the bad side, we loose the
perfect isolation we had before, where only whoever was listening for
something was ever woken up.

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>