libwimaxll: callbacks take "void *" priv pointers
Change the callback model not to require a typed context and make it
completely up to the application.

The reason we had it before using an specific type that could be
wrapped was that we had the callback pre-handler, when de-marshalled
the arguments for the callback, to store in there the result of the
demarshalling.

However, that's kind of silly -- there is nothing the callback can do
in case of error, as it doesn't even get called. So changed to just
print an error message and return to wimaxll_gnl_cb(), which will just
skip the bad netlink message.

As part of this, we can move all the definitions of
'wimaxll_gnl_cb_context' to internal.h.

Note as well that we cannot always set the context passed by
wimaxll_recv(), as that'd mean -EBUSY from the user-set callbacks in
wmx would leak to wimaxll_recv(), when their mission was to force
nl_recvmsgs() to stop processing messages and return to
wimaxll_recv()'s caller inmediately.

So we need to consider -EBUSY as no error but just a code and in that
case, consider that result is "zero".

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>