autoconf: deprecate --with-i2400m, replace w/ --with-linux
The wimax-tools package is not bound to the i2400m driver at all, and
only really needs to link against the Linux kernel's WiMAX stack user
space interface definitions.

So name wise, this option is more congruent.

As well, it now defaults to whatever is installed in /usr/include/linux.

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>