wimaxll_open: if the interface requested is not WiMAX, fail
Due to the changes done during the mainline merge, wimaxll_open()
wasn't really verifying that the interface being requested was a WiMAX
interface. As such, the user would only find out when trying to
execute any function, as it would fail with -ENODEV.

So now, before returning in wimaxll_open(), call rfkill query. If it
fails with -ENODEV, we know the device is not WiMAX.

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>