libwimaxll: wimaxll_recv() doesn't use -ENODATA internally
wimaxll_revc() was returning -ENODATA when an unknown message was
received from the kernel -- this really makes no sense as there is
nothing the client can really do about it.

As well, if no callback was executed, it is considered to be still not
a failure (and thus, don't return -EINPROGRESS).

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>